# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple manga reader app with rate, favorite and comment features",
        "title": "Manga Explorer API",
        "contact": {
            "name": "arcorium",
            "url": "github.com/arcorium",
            "email": "arcorium.l@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/chapters/{chapter_id}": {
            "get": {
                "description": "get specific chapter details with pages associated with it",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga",
                    "chapter"
                ],
                "summary": "Find Chapter Details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "$ref": "#/definitions/dto.ChapterResponse"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "edit specific chapter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga",
                    "chapter"
                ],
                "summary": "Edit Chapter",
                "parameters": [
                    {
                        "description": "chapter edit input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChapterEditInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "delete specific manga chapter and pages associated with it",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga",
                    "chapter"
                ],
                "summary": "Delete Chapter",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/chapters/{chapter_id}/comments": {
            "get": {
                "description": "get all comments from specific chapter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga",
                    "chapter"
                ],
                "summary": "Find Chapter Comment",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/dto.CommentResponse"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "create comment for specific chapter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga",
                    "chapter"
                ],
                "summary": "Create Chapter Comment",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/volumes/{volume_id}": {
            "get": {
                "description": "Get specific volume details with the chapters associated with it",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga",
                    "chapter"
                ],
                "summary": "Find Volume Details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "$ref": "#/definitions/dto.VolumeResponse"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ChapterResponse": {
            "type": "object",
            "properties": {
                "chapter": {
                    "type": "integer"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CommentResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "pages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PageResponse"
                    }
                },
                "title": {
                    "type": "string"
                },
                "total_comment": {
                    "type": "integer"
                },
                "translator": {
                    "$ref": "#/definitions/dto.UserResponse"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorWrapper": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ErrorResponse"
                }
            }
        },
...
        "dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "page": {
                    "$ref": "#/definitions/dto.ResponsePage"
                }
            }
        },
        "dto.SuccessWrapper": {
            "type": "object",
            "properties": {
                "success": {
                    "$ref": "#/definitions/dto.SuccessResponse"
                }
            }
        },
...
    }
}